Crate quill_index

Structs

AspectI

An aspect.

DataI

A data declaration.

DefaultImpls

Tracks the definitions in a project which are marked as “default”.

DefinitionI

A top-level definition, i.e. a def block. TODO: In the future, we will need to add a list of constraints to definitions and data blocks.

EnumI

A enum declaration.

EnumVariant

FileIndex

An index of all top-level items in a file.

ProjectIndex

TypeConstructorI

TypeDeclarationI

A type declaration, e.g. data Bool = True | False.

TypeParameter

UsedFile

Enums

TypeDeclarationTypeI

Either a data declaration, a new type wrapper, a type alias, or a dependent type.

Functions

compute_used_files

Produces a list of all the files (including itself) that are used in this file. project_files is a predicate saying whether the given file is in the project.

index_project

index_single_file